refactor: extract StyleTab from ChartPropertiesPanel
